+++ redwood-layout-view/src/test/kotlin/app/cash/redwood/layout/view/ViewFlexContainerTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.layout
+package app.cash.redwood.layout.view
 
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
@@ -24,7 +24,6 @@ import app.cash.redwood.flexcontainer.AlignItems
 import app.cash.redwood.flexcontainer.FlexDirection
 import app.cash.redwood.flexcontainer.JustifyContent
 import app.cash.redwood.layout.api.Padding
-import app.cash.redwood.layout.view.ViewFlexContainer
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
